backend/requirements.txt
fastapi==0.115.12
uvicorn==0.33.0
sqlalchemy==2.0.36
pydantic-settings==2.8.1
python-multipart==0.0.20
pymupdf==1.24.11
pillow>=7.2.0
httpx==0.28.1
pytest==8.3.5
eval-type-backport==0.3.1
